add event-bus as depeendency and create UserLoggedInEvent class
