+++ src/index.ts
@@ -1,1 +1,6 @@
 export * from './events';
+
+export enum LiberoEventType {
+    serviceStartedIdentifier = 'libero:infra:audit:ServiceStarted',
+    userLoggedInIdentifier = 'libero:audit:user:LoggedIn',
+}
